backend/pyproject.toml
[project]
name = "droneai-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bs4>=0.0.2",
    "fastapi>=0.115.12",
    "geocoder>=1.38.1",
    "httpx>=0.28.1",
    "lxml>=5.4.0",
    "mcp>=1.9.1",
    "openai>=1.79.0",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "uvicorn>=0.34.2",
]
